requirements.txt
appscript==1.3.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
docx2pdf==0.1.8
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
lxml==6.0.2
oauthlib==3.3.1
pdfkit==1.0.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-docx==1.2.0
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wkhtmltopdf==0.2
dotenv==0.9.9 
python-dotenv==1.1.1